+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GPIO: Allow interoperability with other code that initializes the GPIO ISR service (#537)
 - SD card support is no longer behind the `experimental` feature
 - Added 6th TX channel config argument for IDF 5.5.2+.
+- esp32p4 uses vPortExitCriticalMultiCore instead of vPortExitCritical
 
 ## [0.45.2] - 2025-01-15